# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NADITOS API
  description: |
    National Digital Transport Operating System — public HTTP API.

    Every request flows through the API gateway, which:
      - injects an `X-Request-Id` header,
      - verifies the bearer JWT for protected routes,
      - enforces per-route role gates and per-(tenant, route) rate limits.

    Tenant identity is carried in the JWT claim `tid`. Clients may send
    `X-Tenant-Id` for routes that don't require auth (login); for
    authenticated routes the gateway rejects mismatches.
  version: 1.0.0
servers:
  - url: https://api.naditos.example
    description: Production gateway
  - url: http://localhost:8080
    description: Local docker-compose
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /v1/licenses:
    get:
      summary: Search/lookup driver licenses (?number=DL-12345)
      parameters:
        - { name: number, in: query, schema: { type: string } }
      responses: { '200': { description: ok } }
    post:
      summary: Issue a driver license
      responses: { '201': { description: created } }
...
  /v1/licenses/verify:
    post:
      summary: Officer verifies a citizen's QR/NFC bundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties: { token: { type: string } }
      responses: { '200': { description: ok } }
...